Switch from /cache based flashing to (internal) storage based flashing
Copying the file to flash to /cache was an easy and seemingly reliable
solution to give a simple path to the recoveries and have it all working
nicely. Failed to take into account that especially on older devices,
the cache partition is not actually large enough to hold a full 
NIGHTLY ZIP file (not even in compressed format). Though it makes
the situation a bit more messy with the recovery scripts, it does let us
skip a copy stage, which is nice.